Added save to user defined path & display results
Working on displaying both the original and the compressed image
after program is run to show user that the operation was successful.
Currently able to specify if the user would like to overwrite the original
path as well as picking a specific path if the user wants to save to a new place.